requirements.txt
streamlit==1.34.0
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
transformers==4.40.2
requests==2.31.0
python-dotenv==1.0.1
# For Gemini Pro API integration (Google's genai)
google-generativeai==0.4.1
ipwhois